refactor: 重构ssr view输出逻辑

+++ packages/a8k/src/plugins/command-build.ts
@@ -62,7 +62,7 @@ export default class BuildCommand {
           .then(() => hooks.invokePromise('afterBuild', context));
 
         const { ssrConfig } = context.config;
-        if (ssrConfig.entry) {
+        if (ssrConfig.entry && Object.keys(ssrConfig.entry).length) {
           await hooks.invokePromise('beforeSSRBuild', context);
 
           fs.emptyDirSync(ssrConfig.dist);
